+++ src/components/admin/agendor/SyncPanel.tsx
