+++ supabase/config.toml
@@ -2,9 +2,3 @@ project_id = "gikjhasawuhylrpchfqv"
 
 [functions.asaas-webhook]
 verify_jwt = false
-
-[functions.agendor-submit-signup-lead]
-verify_jwt = false
-
-[functions.agendor-sync-all-companies]
-verify_jwt = true
